Use standard Cargo integration test layout
