--- crates/shelterwood/tests/acceptance/mod.rs
@@ -1,8 +1,0 @@
-//! Application-scale acceptance scenarios from `SPEC.md` Appendix C.
-
-register_test_modules!(
-    "tests/acceptance";
-    mod assistant;
-    mod shard_store;
-    mod sidecar;
-);
