+++ crates/shelterwood/tests/acceptance_shard_store.rs
@@ -1,3 +1,5 @@
+mod common;
+
 use std::{
     collections::HashMap,
     sync::{Arc, Mutex},
