+++ crates/shelterwood/tests/common/gates.rs
@@ -137,71 +137,3 @@ impl Drop for DestructorBlocker {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::{
-        future::Future,
-        sync::atomic::Ordering,
-        task::{Context, Waker},
-    };
-
-    use super::ReleaseGate;
-
-    #[tokio::test]
-    async fn release_gate_stores_one_permit() {
-        let gate = ReleaseGate::default();
-        gate.release();
-        gate.wait().await;
-    }
-
-    #[test]
-    #[should_panic(expected = "cannot run more than one release ahead of its waiters")]
-    fn release_gate_rejects_a_release_no_waiter_asked_for() {
-        let gate = ReleaseGate::default();
-        gate.release();
-        gate.release();
-    }
-
-    #[tokio::test]
-    async fn release_gate_serves_every_wait_that_has_begun() {
-        let gate = ReleaseGate::default();
-        let first = tokio::spawn({
-            let gate = gate.clone();
-            async move { gate.wait().await }
-        });
-        let second = tokio::spawn({
-            let gate = gate.clone();
-            async move { gate.wait().await }
-        });
-        while gate.0.waits.load(Ordering::SeqCst) < 2 {
-            tokio::task::yield_now().await;
-        }
-
-        // Back-to-back releases for two already-parked waiters are ordinary
-        // usage: the check counts demand, not whichever permits the semaphore
-        // happens to be storing at this instant.
-        gate.release();
-        gate.release();
-
-        first.await.expect("first waiter joins");
-        second.await.expect("second waiter joins");
-    }
-
-    #[tokio::test]
-    async fn cancelling_a_wait_returns_its_release_to_the_gate() {
-        let gate = ReleaseGate::default();
-        let mut waiting = Box::pin(gate.wait());
-        let mut context = Context::from_waker(Waker::noop());
-        assert!(waiting.as_mut().poll(&mut context).is_pending());
-
-        gate.release();
-        // Tokio hands the permit to the parked waiter and takes it back when
-        // that future is dropped. The wait still counts as demand, so the
-        // replacement release is not an over-release.
-        drop(waiting);
-        gate.release();
-
-        gate.wait().await;
-    }
-}
