+++ crates/shelterwood/tests/observation.rs
@@ -1,3 +1,5 @@
+mod common;
+
 use std::{
     collections::HashMap,
     sync::{
